Fixes #2
- Omintruck script is only used if the package is not yet installed

+++ recipes/chef.rb
@@ -6,6 +6,7 @@ execute 'Install chefdk with omnitruck' do
   command 'curl -L https://omnitruck.chef.io/install.sh | sudo bash -s -- -P chefdk'
   notifies :delete, 'file[Delete omnitruck install script]', :before
   notifies :delete, 'file[Delete omnitruck install script]', :immediately
+  not_if 'dpkg -l chefdk'
 end
 
 file 'Delete omnitruck install script' do
